SurfaceFlinger_hwc1: fix state recognition of color calibration matrix
Previously, the state of the color calibration matrix was decided by the framework which sent an integer to determine it (non-zero: enable, zero: disabled).

This caused the flinger to wrongly keep the matrix enabled (and the HWC disabled) even though the user disabled the color calibration utility.

Correct this by also checking the actual sum the the matrix, with 4 being corresponding value for rgb(255, 255, 255), meaning color calibration is disabled.